feat(Invoice_MakerApp.swift): add ContactStoreManager to manage contact store feat(Customer.swift): add initializer to create Customer from CNContact feat(CustomersContactView.swift): add new view to handle contact access and fetching feat(CustomersListItemView.swift): add new view to display customer list item feat(CustomersView.swift): add search functionality and contact import feature
This commit introduces several features to improve the user experience and functionality of the app. The ContactStoreManager is added to manage the contact store, and a new initializer is added to the Customer class to create a Customer instance from a CNContact. A new view, CustomersContactView, is added to handle contact access and fetching. Another new view, CustomersListItemView, is added to display each customer as a list item. Finally, search functionality and the ability to import contacts are added to the CustomersView.